Compile errors. Sector: int, char -> bytes van 60 bytes/Sector 48 bytes/Sector, 6000 bytes winst voor 500 sectoren. Robot rijdt, maar rijdt tegen muren, ook al weet hij dat er één is?!?
MemoryTesten verplaatst

+++ src/java/penoplatinum/grid/SimpleGrid.java
@@ -288,7 +288,7 @@ public class SimpleGrid implements Grid {
     }
     if (s.getTagCode() != -1) {
       thisSector.setTagCode(s.getTagCode());
-      thisSector.setTagBearing((s.getTagBearing() + rotation) % 4);
+      thisSector.setTagBearing((byte)((s.getTagBearing() + rotation) % 4));
     }
 
   }
